Public API

Umatrix.bestMatch — Function

bestMatch(dataPoint::AbstractVector{Float64}, weights::EsomWeights{Float64})

Search for the best matching ESOM neuron for the given data point.

Umatrix.esomTrain — Function

esomTrain(data::AbstractMatrix{Float64})

Train an ESOM for the given data set.

Uses on-line learning per default. (Batch is not yet implemented)

Umatrix.esomTrainOnline — Function

esomTrainOnline(data::AbstractMatrix{Float64})

Train an ESOM on-line for the given data set.

Umatrix.plot_matrix! — Method

plot_matrix(m; <keyword arguments>)

Visualize the given U-, P- or U*-matrix.

Arguments

projection: A projection of data points to ESOM neuron to show on the map.

classes: The class of each data point. Used to color the dots for the projection.

normalize: Whether to normalize the given matrix. Default: true.

colorStyle: Either :umatrix or :pmatrix. Default: :umatrix.

Umatrix.plot_matrix! — Method

plot_matrix(m; <keyword arguments>)

Visualize the given U-, P- or U*-matrix.

Arguments

projection: A projection of data points to ESOM neuron to show on the map.

classes: The class of each data point. Used to color the dots for the projection.

normalize: Whether to normalize the given matrix. Default: true.

colorStyle: Either :umatrix or :pmatrix. Default: :umatrix.

Umatrix.plot_matrix — Method

plot_matrix(m; <keyword arguments>)

Visualize the given U-, P- or U*-matrix.

Arguments

projection: A projection of data points to ESOM neuron to show on the map.

classes: The class of each data point. Used to color the dots for the projection.

normalize: Whether to normalize the given matrix. Default: true.

colorStyle: Either :umatrix or :pmatrix. Default: :umatrix.

Umatrix.pmatrix — Function

pmatrix(data::AbstractMatrix{Float64}, weights::EsomWeights{Float64}; radius = nothing)

Generate a P-matrix for the given data and ESOM weights.

If no pareto radius is given, a suitable one is estimated.

Umatrix.projection — Function

projection(data::AbstractMatrix{Float64}, weights::EsomWeights{Float64})

Generate a projection from each data point to the best matching ESOM neuron.

Umatrix.umatrix — Function

umatrix(weights::EsomWeights{Float64})

Generate a U-matrix for the given ESOM weights.

Umatrix.ustarmatrix — Function

ustarmatrix(data::AbstractMatrix{Float64}, weights::EsomWeights{Float64})

Generate a U*-matrix for the given data and ESOM weights.

Umatrix.ustarmatrix — Function

ustarmatrix(um::AbstractMatrix{Float64}, pm::AbstractMatrix{Int})

Generate a U*-matrix from the given U- and P-matrices.

Internal API

Umatrix.coolDown — Method

coolDown(::Val{:leadInOut}, (start, stop), steps; leadIn = 0.1, leadOut = 0.95)

Return a linear cooldown function with lead-in and lead-out.

The range start - stop is devided into equidistant steps. If the given step is below the leadIn quantile, start is returned; if it's above the leadOut quantile, stop is returned.

Umatrix.coolDown — Method

coolDown(::Val{:linear}, (start, stop), steps)

Return a linear cooldown function.

The range start - stop is devided into equidistant steps.

Umatrix.correctCoords — Function

correctCoords(coords::AbstractVector{CartesianIndex{2}})

Remove (if the lattice is a plane) or wrap (if the lattice is a toroid) coordinates outside of the lattice size.

Umatrix.directNeighbours — Function

directNeighbours(ind::CartesianIndex{2})

Return the eight direct neighbours of index ind.

Umatrix.esomInit — Function

esomInit(data::AbstractMatrix{Float64}, settings::Settings = defaultSettings)

Initilize an ESOM with values based on settings.initMethod.

Umatrix.esomTrainEpoch! — Function

esomTrainEpoch!(data, weights, radius, learningRate)

Train the ESOM for a single epoch.

This function mutates weights

Umatrix.esomTrainOnline! — Function

esomTrainOnline!(data::AbstractMatrix{Float64}, weights::EsomWeights{Float64})

Train the ESOM weights on-line for the given data set.

This function mutates weights

Umatrix.esomTrainStep! — Function

esomTrainStep!(dataPoint, weights, radius, learningRate)

Train the ESOM with a single data point.

This function mutates weights

Umatrix.initMethod — Function

initMethod(::Val{:uniform_mean_std})

Returns an initializer function.

The values are uniformly distributed between $[μ-2σ, μ+2σ]$ of the data column.

Umatrix.initMethod — Function

initMethod(::Val{:normal_mean_std})

Returns an initializer function.

The values are Normal distributed with the same mean and std as the data column.

Umatrix.initMethod — Function

initMethod(::Val{:uniform_min_max})

Returns an initializer function.

The values are uniformly distributed over $[\min(c), \max(c)]$ of the data column.

Umatrix.initMethod — Function

initMethod(::Val{:zeros})

Returns an initializer function.

The values are set to 0.

Umatrix.latticeDistance — Function

latticeDistance(a::CartesianIndex{2}, b::CartesianIndex{2},

Euclidean distance on the lattice.

Umatrix.neighbourhood — Function

neighbourhoodOffsets(radius::Float64)

Return every (unique) CartesianIndex{2} in a radius-sized circle around index.

Umatrix.neighbourhoodKernel — Method

neighbourhoodKernel(::Val{:cone}, radius::Float64)

Returns a linear neighbourhood decay function.

Umatrix.neighbourhoodKernel — Method

neighbourhoodKernel(::Val{:gauss}, radius::Float64)

Returns a gaussian neighbourhood decay function.

Umatrix.neighbourhoodKernel — Method

neighbourhoodKernel(::Val{:mexhat}, radius::Float64)

Returns a "Mexican hat" (Ricker wavelet) neighbourhood decay function.

Umatrix.neighbourhoodOffsets — Method

neighbourhoodOffsets(radius::Float64)

Return every CartesianIndex{2} in a radius-sized circle around (0,0).

Umatrix.removeCoordsOutsideBounds — Function

removeCoordsOutsideBounds(coords::AbstractVector{CartesianIndex{2}})

Remove coordinates outside of the lattice size.

Umatrix.shiftToHighestDensity — Function

shiftToHighestDensity(data::AbstractMatrix{Float64}, weights::EsomWeights{Float64})

Shift the ESOM so the point of highest density is centered.

Umatrix.shiftWeights — Function

shiftWeights(weights::EsomWeights{Float64}, pos::CartesianIndex{2})

Shift the ESOM so the given pos is in the middle.

This makes only sense for a toroidal map.

Umatrix.wrapCoordsOnToroid — Function

wrapCoordsOnToroid(coords::AbstractVector{CartesianIndex{2}},

Wrap coordinates outside of the lattice size around the torus.